refactor: Restructure application to use blueprints for better organization
- Created ``__init__.py`` and routes `.py` files
